Resizable NavBar Widgets [1/2] WIP
Primarily this creates a longpress event on the widget
label to allow dragging up/down to resize the widget.
It also makes a lot of changes on how we store the widget size,
so a ROMControl change is needed as well.

Smooths out the resizing of widgets.
also provides a shortcut that can be used to toggle widgets.
Renamed the 'SystemUI' label to 'AOKP Actions' for shortcut picking

Change-Id: I753e988e86eef3707ab647383f28e880a81607f8
Signed-off-by: Zaphod <mwilson@acesappraisal.com>