Mass cleanup of navigation bar background transparency
included in this is stuff picked from romans aokp commits for navbar transparency.
this fixes the issues some launchers have with not showing the wallpaper behind the navbar fully.

moved the navbar bg color/alpha from phonestatusbar to navigationbarview improving the way it gets set and removing the need for a systemui restart (change is reflected in baked settings)