Revert: "volume rocker wake: add volume keys to isWakeKey KeyEvents"
This commit is the mysterious reason for the "always-on" volume rocker wake when ambient display was on (at least in my testing).
Searching various ROM commits, I found that this commit was needed in 5.0.x ROMs and it was working correctly, and the bug is only happening in 5.1 ROMs.

Reverting it fixes the bug.